Vendor the GenAI e2e Spring AI application and split out the mock LLM server
The virtual-genai case built its Spring AI application at test time by cloning
spring-projects/spring-ai-examples and running Maven inside the image build. That
resolved spring-ai:2.0.0-SNAPSHOT, a line abandoned the day before the 2.0.0 GA
whose older builds have since been pruned, so the fixture was pinned to nothing
released and could no longer be rebuilt from the same bytes.

Replace it with an in-repo e2e-spring-ai-service module on the released Spring AI
2.0.1, and move the mock LLM endpoints out of e2e-service-provider into a dedicated
e2e-mock-llm-server that all three GenAI cases share -- e2e-service-provider was
only ever started in those cases to host that controller.

Spring AI requires Java 17 and Spring Boot 4, while the e2e reactor targets Java 8
and is compiled under JDK 11 on the e2e-test-java-versions leg, so the new module
sits behind a jdk-17 profile and the jobs that build it now pin JDK 17 explicitly.
