--- test/e2e-v2/cases/virtual-genai/Dockerfile.provider
@@ -1,44 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-ARG SW_AGENT_JDK_VERSION
-ARG SW_AGENT_JAVA_COMMIT
-
-FROM eclipse-temurin:17-jdk AS builder
-
-RUN apt-get update && apt-get install -y git maven && rm -rf /var/lib/apt/lists/*
-
-WORKDIR /source
-RUN git clone https://github.com/spring-projects/spring-ai-examples.git \
-    && cd spring-ai-examples \
-    && git checkout 2a6088db3d18d5fa6fc208b12adf1172d22f77fd
-
-WORKDIR /source/spring-ai-examples/misc/openai-streaming-response
-RUN mvn clean package -DskipTests
-
-RUN find target/ -maxdepth 1 -name "*.jar" ! -name "*-plain.jar" -exec cp {} /app.jar \;
-
-FROM ghcr.io/apache/skywalking-java/skywalking-java:${SW_AGENT_JAVA_COMMIT}-java${SW_AGENT_JDK_VERSION}
-
-WORKDIR /skywalking
-
-VOLUME /services
-
-COPY --from=builder /app.jar /services/app.jar
-
-EXPOSE 8080
-
-CMD ["sh", "-c", "java -jar /services/app.jar"]
-
