+++ test/e2e-v2/java-test-service/e2e-mock-llm-server/src/main/java/org/apache/skywalking/e2e/mockllm/LLMMockController.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.skywalking.e2e.controller;
+package org.apache.skywalking.e2e.mockllm;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
