+++ test/e2e-v2/java-test-service/e2e-spring-ai-service/src/main/java/org/apache/skywalking/e2e/springai/SpringAIServiceApplication.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.skywalking.e2e.springai;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * A Spring AI client driven by the GenAI e2e cases. It exists so the Java agent's spring-ai plugin
+ * has something real to instrument; the LLM it calls is e2e-mock-llm-server, not a live provider.
+ */
+@SpringBootApplication
+public class SpringAIServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(SpringAIServiceApplication.class, args);
+    }
+}
