+++ test/e2e-v2/java-test-service/e2e-spring-ai-service/src/main/java/org/apache/skywalking/e2e/springai/StreamingChatController.java
@@ -1,0 +1,52 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.skywalking.e2e.springai;
+
+import org.springframework.ai.chat.messages.UserMessage;
+import org.springframework.ai.chat.model.ChatResponse;
+import org.springframework.ai.chat.prompt.Prompt;
+import org.springframework.ai.openai.OpenAiChatModel;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+
+@RestController
+@RequestMapping("/ai")
+public class StreamingChatController {
+    private final OpenAiChatModel chatModel;
+
+    public StreamingChatController(final OpenAiChatModel chatModel) {
+        this.chatModel = chatModel;
+    }
+
+    /**
+     * Streams a chat completion. The e2e trigger calls this; the span the agent produces for the
+     * underlying {@code ChatModel.stream} call is what the GenAI assertions are written against.
+     *
+     * @param message the prompt to send to the model
+     * @return the model's response chunks, as server-sent events
+     */
+    @GetMapping(value = "/generateStream", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
+    public Flux<ChatResponse> generateStream(@RequestParam(defaultValue = "Tell me a joke") final String message) {
+        return chatModel.stream(new Prompt(new UserMessage(message)));
+    }
+}
