+++ test/e2e-v2/java-test-service/e2e-spring-ai-service/src/main/resources/application.yml
@@ -1,0 +1,22 @@
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+spring:
+  main:
+    banner-mode: 'off'
+  ai:
+    openai:
+      api-key: ${OPENAI_API_KEY:e2e-api-key}
+      base-url: ${SPRING_AI_OPENAI_BASE_URL:http://localhost:9090/llm/v1}
